==> 01-preprocess-data.ipynb <==
import pandas as pd

df_ratings = pd.read_csv('data/original/ratings.csv')
df_personality = pd.read_csv('data/original/personality-data.csv')

Preprocess Ratings dataset

df_ratings.head()

# Remove unnecessary columns
df_ratings = df_ratings.drop(columns=[' tstamp '])

# Clean up column names
df_ratings.rename(columns={"useri": "raw_user_id", " movie_id": "movie_id", " rating": "rating"}, inplace=True)

# Map raw_user_id to integer values that is suitable for PyTorch
user_mapping = { user_id: i for i, user_id in enumerate(df_ratings['raw_user_id'].unique())}
movie_mapping = { movie_id: i for i, movie_id in enumerate(df_ratings['movie_id'].unique())}
df_ratings['user_id'] = df_ratings['raw_user_id'].map(user_mapping)
df_ratings['movie_id'] = df_ratings['movie_id'].map(movie_mapping)

# Drop raw_user_id column
df_ratings = df_ratings.drop(columns=['raw_user_id'])

# Divide the dataset into train and test sets
from sklearn.model_selection import train_test_split

df_ratings_train, df_ratings_test = train_test_split(df_ratings, test_size=0.2, random_state=42)

# Remove rows in test set where user_id is not in train set
df_ratings_test = df_ratings_test[df_ratings_test['user_id'].isin(df_ratings_train['user_id'])]

# Save the modified ratings and train and test sets as CSV files in data/preprocessed
df_ratings.to_csv('data/preprocessed/ratings.csv', index=False)
df_ratings_train.to_csv('data/preprocessed/ratings-train.csv', index=False)
df_ratings_test.to_csv('data/preprocessed/ratings-test.csv', index=False)

Preprocess Personality dataset

# Extract columns userid, openness, agreeableness, emotional_stability, conscientiousness, extraversion from df_personality
df_personality = df_personality[['userid', ' openness', ' agreeableness', ' emotional_stability', ' conscientiousness', ' extraversion']]
df_personality.head()

df_personality.columns

# Rename columns to match the OCEAN model names
df_personality.rename(columns={
	"userid": "raw_user_id",
	" openness": "openness", 
	" conscientiousness": "conscientiousness",
	" extraversion": "extraversion",
	" emotional_stability": "neuroticism", 
	" agreeableness": "agreeableness",
	}, 
	inplace=True)

df_personality.head()

# Map raw_user_id to integer values used in df_ratings
df_personality['user_id'] = df_personality['raw_user_id'].map(user_mapping)
df_personality.head()

# Drop raw_user_id and save the dataset as CSV file in data/preprocessed
df_personality = df_personality.drop(columns=['raw_user_id'])
df_personality.to_csv('data/preprocessed/personality.csv', index=False)

# Merge personality into train and test dataset
df_ratings_train = pd.merge(df_ratings_train, df_personality, on='user_id')
df_ratings_test = pd.merge(df_ratings_test, df_personality, on='user_id')

# Save them
df_ratings_train.to_csv('data/preprocessed/ratings-train.csv', index=False)
df_ratings_test.to_csv('data/preprocessed/ratings-test.csv', index=False)

Personality test with user_id in rating-test

df_personality = pd.read_csv('data/preprocessed/personality.csv')
df_ratings_test = pd.read_csv('data/preprocessed/ratings-test.csv')

# Remove rows from df_personality which have user_id which are not present in df_rating_test
df_personality = df_personality[df_personality['user_id'].isin(df_ratings_test['user_id'])]
df_personality.to_csv('data/preprocessed/personality-test.csv', index=False)
